requirements.txt
allure-pytest~=2.9.45
jsonpath~=0.82
requests~=2.26.0
Jinja2~=3.0.3
loguru~=0.6.0
pika~=1.2.1
retry~=0.9.2
PyMySQL~=1.0.2
paramiko~=2.8.1
celery~=5.2.7
redis~=4.3.3
pytest~=6.2.5
nacos-sdk-python~=0.1.8
PyYAML~=6.0
pycryptodome~=3.15.0
gmssl~=3.2.1
cacheout~=0.14.1
fastapi~=0.85.0
uvicorn~=0.18.3